Fix a number of MSVC warnings introduced by RFC31
git-svn-id: https://svn.osgeo.org/gdal/trunk@28382 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/dxf/ogrdxfwriterlayer.cpp
@@ -212,7 +212,7 @@ OGRErr OGRDXFWriterLayer::WriteCore( OGRFeature *poFeature )
 /*      Also, for reasons I don't understand these ids seem to have     */
 /*      to start somewhere around 0x50 hex (80 decimal).                */
 /* -------------------------------------------------------------------- */
-    poFeature->SetFID( poDS->WriteEntityID(fp,poFeature->GetFID()) );
+    poFeature->SetFID( poDS->WriteEntityID(fp,(int)poFeature->GetFID()) );
 
 /* -------------------------------------------------------------------- */
 /*      For now we assign everything to the default layer - layer       */
